+++ lib/src/core/enums/todo_operation_type.dart
@@ -1,0 +1,4 @@
+enum TodoOperationType {
+  create,
+  edit,
+}
